+++ unlicense/resources/frida.js
@@ -1,6 +1,6 @@
 "use strict";
 
-let allocated_buffers = [];
+let allocatedBuffers = [];
 
 function log(message) {
     console.log(`frida-agent: ${message}`);
@@ -12,35 +12,35 @@ function walk_back_stack_for_oep(context, module) {
         return null;
     }
 
-    const original_text_section = Process.findRangeByAddress(backtrace[0]);
-    if (original_text_section == null) {
+    const originalTextSection = Process.findRangeByAddress(backtrace[0]);
+    if (originalTextSection == null) {
         return null;
     }
 
-    const module_start = module.base;
-    const module_end = module.base.add(module.size);
-    if (module_start.compare(original_text_section.base) > 0 || module_end.compare(original_text_section.base) <= 0) {
+    const moduleStart = module.base;
+    const moduleEnd = module.base.add(module.size);
+    if (moduleStart.compare(originalTextSection.base) > 0 || moduleEnd.compare(originalTextSection.base) <= 0) {
         return null;
     }
 
-    let oep_candidate = null;
-    const text_section_start = original_text_section.base;
-    const text_section_end = original_text_section.base.add(original_text_section.size);
+    let oepCandidate = null;
+    const textSectionStart = originalTextSection.base;
+    const textSectionEnd = originalTextSection.base.add(originalTextSection.size);
     backtrace.forEach(addr => {
-        if (text_section_start.compare(addr) <= 0 && text_section_end.compare(addr) > 0) {
-            oep_candidate = addr;
+        if (textSectionStart.compare(addr) <= 0 && textSectionEnd.compare(addr) > 0) {
+            oepCandidate = addr;
         }
     });
-    return oep_candidate;
+    return oepCandidate;
 }
 
-function compute_real_oep(oep_candidate) {
+function compute_real_oep(oepCandidate) {
     if (Process.arch == "ia32") {
         // Note: This assumes the OEP looks like this (MSVC) and `oep_candidate`
         // is located after the call:
         //   call __security_init_cookie
         //   jmp __tmainCRTStartup
-        return oep_candidate.sub(5);
+        return oepCandidate.sub(5);
     } else if (Process.arch == "x64") {
         // Note: This assumes the OEP looks like this (MSVC) and `oep_candidate`
         // is located after the call:
@@ -48,33 +48,63 @@ function compute_real_oep(oep_candidate) {
         //   call __security_init_cookie
         //   add rsp, 0x28
         //   jmp __tmainCRTStartup
-        return oep_candidate.sub(9);
+        return oepCandidate.sub(9);
     } else {
         // FIXME
-        return oep_candidate;
+        return oepCandidate;
     }
 }
 
 // Define available RPCs
 rpc.exports = {
-    setupOepTracing: function (module_name) {
-        const dumped_module = Process.findModuleByName(module_name);
-        if (dumped_module == null) {
+    setupOepTracing: function (moduleName) {
+        const dumpedModule = Process.findModuleByName(moduleName);
+        if (dumpedModule == null) {
             log('Invalid module specified');
             return;
         }
+        log(`Setting up OEP tracing for "${moduleName}"`);
 
-        log(`Setting up OEP tracing for "${module_name}"`);
-        const RtlQueryPerformanceCounter = Module.findExportByName('ntdll', 'RtlQueryPerformanceCounter')
-        Interceptor.attach(RtlQueryPerformanceCounter, {
-            onEnter: function (_args) {
-                let oep_candidate = walk_back_stack_for_oep(this.context, dumped_module);
-                if (oep_candidate != null) {
-                    oep_candidate = compute_real_oep(oep_candidate);
-                    log(`Possible OEP (thread #${this.threadId}): ${oep_candidate}`);
-                    send({ 'event': 'oep_reached', 'OEP': oep_candidate, 'BASE': dumped_module.base })
-                    let sync_op = recv('block_on_oep', function (_value) { });
-                    sync_op.wait();
+        // Hook `ntdll.NtMapViewOfSection` as a mean to get called when new PE
+        // images are loaded.
+        const ntMapViewOfSection = Module.findExportByName('ntdll', 'NtMapViewOfSection');
+        let queryPerformanceCounterHooked = false;
+        let corExeMainHooked = false;
+        Interceptor.attach(ntMapViewOfSection, {
+            onLeave: function (_args) {
+                // Hook `kernel32.QueryPerformanceCounter` if present
+                // This is used to detect the CRT entry point for EXEs compiled
+                // from C, C++ (or Deplhi) 
+                const queryPerformanceCounter = Module.findExportByName('kernel32', 'QueryPerformanceCounter');
+                if (queryPerformanceCounter != null && !queryPerformanceCounterHooked) {
+                    Interceptor.attach(queryPerformanceCounter, {
+                        onEnter: function (_args) {
+                            let oepCandidate = walk_back_stack_for_oep(this.context, dumpedModule);
+                            if (oepCandidate != null) {
+                                oepCandidate = compute_real_oep(oepCandidate);
+                                log(`Potential OEP (thread #${this.threadId}): ${oepCandidate}`);
+                                send({ 'event': 'oep_reached', 'OEP': oepCandidate, 'BASE': dumpedModule.base, 'DOTNET': false })
+                                let sync_op = recv('block_on_oep', function (_value) { });
+                                sync_op.wait();
+                            }
+                        }
+                    });
+                    queryPerformanceCounterHooked = true;
+                }
+                // Hook `clr.InitializeFusion` if present.
+                // This is used to detect a good point during the CLR's
+                // initialization, to dump .NET EXE assemblies
+                const corExeMain = Module.findExportByName('clr', 'InitializeFusion');
+                if (corExeMain != null && !corExeMainHooked) {
+                    Interceptor.attach(corExeMain, {
+                        onEnter: function (_args) {
+                            log(`Potential .NET assembly entry (thread #${this.threadId})`);
+                            send({ 'event': 'oep_reached', 'OEP': '0', 'BASE': dumpedModule.base, 'DOTNET': true })
+                            let sync_op = recv('block_on_oep', function (_value) { });
+                            sync_op.wait();
+                        }
+                    });
+                    corExeMainHooked = true;
                 }
             }
         });
@@ -92,17 +122,17 @@ rpc.exports = {
     },
     enumerateModules: function () {
         const modules = Process.enumerateModules();
-        let module_names = [];
+        let moduleNames = [];
         modules.forEach(module => {
-            module_names = module_names.concat(module.name);
+            moduleNames = moduleNames.concat(module.name);
         });
-        return module_names;
+        return moduleNames;
     },
-    enumerateModuleRanges: function (module_name) {
+    enumerateModuleRanges: function (moduleName) {
         let ranges = Process.enumerateRangesSync("r--");
         return ranges.filter(range => {
             const module = Process.findModuleByAddress(range.base);
-            return module != null && module.name.toUpperCase() == module_name.toUpperCase();
+            return module != null && module.name.toUpperCase() == moduleName.toUpperCase();
         });
     },
     enumerateExportedFunctions: function () {
@@ -114,9 +144,9 @@ rpc.exports = {
         return exports;
     },
     allocateProcessMemory: function (size, near) {
-        let size_rounded = size + (Process.pageSize - size % Process.pageSize);
-        let addr = Memory.alloc(size_rounded, { near: ptr(near), maxDistance: 0xff000000 });
-        allocated_buffers.push(addr)
+        let sizeRounded = size + (Process.pageSize - size % Process.pageSize);
+        let addr = Memory.alloc(sizeRounded, { near: ptr(near), maxDistance: 0xff000000 });
+        allocatedBuffers.push(addr)
         return addr;
     },
     queryMemoryProtection: function (address) {